重写Post界面
重写Post界面，去除翻页框架
