+++ WBSBlog/Home/Controllers/WBSBasePostViewController.h
@@ -1,0 +1,13 @@
+//
+//  WBSBasePostViewController.h
+//  WBSBlog
+//
+//  Created by Weberson on 2017/2/13.
+//  Copyright © 2017年 Weberson. All rights reserved.
+//
+
+#import "WBSBaseViewController.h"
+
+@interface WBSBasePostViewController : WBSBaseViewController
+
+@end
