+++ WBSBlog/Home/Controllers/WBSBasePostViewController.m
@@ -1,0 +1,37 @@
+//
+//  WBSBasePostViewController.m
+//  WBSBlog
+//
+//  Created by Weberson on 2017/2/13.
+//  Copyright © 2017年 Weberson. All rights reserved.
+//
+
+#import "WBSBasePostViewController.h"
+
+@interface WBSBasePostViewController ()
+
+@end
+
+@implementation WBSBasePostViewController
+
+- (void)viewDidLoad {
+    [super viewDidLoad];
+    // Do any additional setup after loading the view.
+}
+
+- (void)didReceiveMemoryWarning {
+    [super didReceiveMemoryWarning];
+    // Dispose of any resources that can be recreated.
+}
+
+/*
+#pragma mark - Navigation
+
+// In a storyboard-based application, you will often want to do a little preparation before navigation
+- (void)prepareForSegue:(UIStoryboardSegue *)segue sender:(id)sender {
+    // Get the new view controller using [segue destinationViewController].
+    // Pass the selected object to the new view controller.
+}
+*/
+
+@end
