+++ WBSBlog/Home/Controllers/WBSHomePostViewController.h
@@ -2,51 +2,16 @@
 //  WBSHomePostViewController.h
 //  WBSBlog
 //
-//  Created by Weberson on 16/7/20.
-//  Copyright © 2016年 Weberson. All rights reserved.
+//  Created by Weberson on 2017/2/13.
+//  Copyright © 2017年 Weberson. All rights reserved.
 //
 
-#import <UIKit/UIKit.h>
 #import "WBSBasePostViewController.h"
-
-/**
- *  文章类型枚举
- */
-typedef NS_ENUM(NSUInteger, PostType){
-    /**
-     *  文章
-     */
-    PostTypePost,
-    /**
-     *  页面
-     */
-    PostTypePage,
-};
-
-
-/**
- *  API 类型枚举 15-08-04 by weberson
- */
-typedef NS_ENUM(NSInteger,APIType){
-    /**
-     *  JSON API
-     */
-    APITypeJSON,
-    /**
-     *  XMLRPC API
-     */
-    APITypeXMLRPC,
-    /**
-     *  Http API
-     */
-    APITypeHttp
-};
-
 
 /**
  *  文章结果类型枚举
  */
-typedef NS_ENUM(NSUInteger, PostResultType){
+typedef NS_ENUM(NSUInteger, PostViewType){
     /**
      *  最近文章
      */
@@ -67,57 +32,10 @@ typedef NS_ENUM(NSUInteger, PostResultType){
 
 @interface WBSHomePostViewController : WBSBasePostViewController
 
-/**
- *  API类型
- */
-@property (nonatomic,assign) APIType apiType;
-/**
- *  文章类型
- */
-@property (nonatomic,assign) PostType postType;
-/**
- *  文章结果类型
- */
-@property (nonatomic,assign) PostResultType postResultType;
-/**
- *  文章列表数据
- */
-@property (nonatomic,strong) NSArray *posts;
-/**
- *  是否搜索
- */
-@property (nonatomic,assign)  BOOL isSearch;
-/**
- *  搜索框
- */
-@property (nonatomic,strong)  UISearchBar *searchBar;
-/**
- *  搜索关键字
- */
-@property (nonatomic,strong) NSString *searchString;
-/**
- *  分类Id
- */
-@property (nonatomic,assign) NSInteger categoryId;
-/**
- *  标签ID
- */
-@property (nonatomic,assign)  NSInteger tagId;
-/**
- *  文章搜索控制器
- */
-@property  (nonatomic,strong) UISearchController *postSearchController;
-/**
- *  下拉分类
- */
-@property (nonatomic, strong) UIButton *titleButton;
-/**
- *  根据文章类型初始化
- *
- *  @param type 文章类型
- *
- *  @return 当前对象
- */
-- (instancetype)initWithPostType:(PostType)type;
 
 @end
+
+
+
+
+
