+++ WBSBlog/Home/View/WBSPostTableView.h
@@ -1,0 +1,13 @@
+//
+//  WBSPostTableView.h
+//  WBSBlog
+//
+//  Created by Weberson on 2017/2/13.
+//  Copyright © 2017年 Weberson. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface WBSPostTableView : UITableView
+
+@end
