+++ WBSBlog/Home/View/WBSPostTableView.m
@@ -1,0 +1,88 @@
+//
+//  WBSPostTableView.m
+//  WBSBlog
+//
+//  Created by Weberson on 2017/2/13.
+//  Copyright © 2017年 Weberson. All rights reserved.
+//
+
+#import "WBSPostTableView.h"
+
+static NSString *PostCellID = @"PostCellID";
+
+@interface WBSPostTableView ()
+
+@property (nonatomic, strong) NSArray *dataArray;  //!< <#属性注释#>
+
+@end
+
+
+@implementation WBSPostTableView
+
+#pragma mark - 数据源方法
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+    return self.dataArray.count;
+}
+
+- (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
+    
+    UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:PostCellID];
+    
+    // 设置 Cell...
+    
+    return cell;
+}
+
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+{
+
+    return 0;
+}
+
+
+// Override to support conditional editing of the table view.
+- (BOOL)tableView:(UITableView *)tableView canEditRowAtIndexPath:(NSIndexPath *)indexPath {
+    // Return NO if you do not want the specified item to be editable.
+    return YES;
+}
+
+// Override to support editing the table view.
+- (void)tableView:(UITableView *)tableView commitEditingStyle:(UITableViewCellEditingStyle)editingStyle forRowAtIndexPath:(NSIndexPath *)indexPath {
+
+}
+
+
+- (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
+{
+
+
+}
+
+
+
+
+
+
+
+
+
+
+
+@end
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
