--- WBSBlog/Main/SwipableViewController/WBSSwipableViewController.m
@@ -1,112 +1,0 @@
-//
-//  WBSSwipableViewController.m
-//  WBSBlog
-//
-//  Created by Weberson on 16/7/20.
-//  Copyright © 2016年 Weberson. All rights reserved.
-//
-
-#import "WBSSwipableViewController.h"
-#import "UIColor+Config.h"
-
-@interface WBSSwipableViewController ()
-
-@end
-
-@implementation WBSSwipableViewController
-
-
-- (instancetype)initWithTitle:(NSString *)title andSubTitles:(NSArray *)subTitles andControllers:(NSArray *)controllers
-{
-    return [self initWithTitle:title andSubTitles:subTitles andControllers:controllers underTabbar:NO];
-}
-
-- (instancetype)initWithTitle:(NSString *)title andSubTitles:(NSArray *)subTitles andControllers:(NSArray *)controllers underTabbar:(BOOL)underTabbar
-{
-    self = [super init];
-  
-    if (self) {
-        self.edgesForExtendedLayout = UIRectEdgeNone;
-        self.navigationController.navigationBar.translucent = NO;
-        self.view.autoresizingMask = UIViewAutoresizingFlexibleHeight | UIViewAutoresizingFlexibleWidth;
-        
-        if (title) {self.title = title;}
-                
-        //默认不显示TitleBar
-        CGFloat titleBarHeight = 0;
-        //当传递了subTitles时才去指定TitleBar
-        if (subTitles) {
-            titleBarHeight = 36;
-            _titleBar = [[WBSTitleBarView alloc] initWithFrame:CGRectMake(0, 0, self.view.bounds.size.width, titleBarHeight) andTitles:subTitles];
-            _titleBar.backgroundColor = [UIColor clearColor];
-            [self.view addSubview:_titleBar];
-
-        }
-        
-        _viewPager = [[WBSHorizonalTableViewController alloc] initWithViewControllers:controllers];
-        
-        CGFloat height = self.view.bounds.size.height - titleBarHeight - 64 - (underTabbar ? 49 : 0);
-        _viewPager.view.frame = CGRectMake(0, titleBarHeight, self.view.bounds.size.width, height);
-        
-        [self addChildViewController:self.viewPager];
-        [self.view addSubview:_viewPager.view];
-
-        
-        __weak WBSTitleBarView *weakTitleBar = _titleBar;
-        __weak WBSHorizonalTableViewController *weakViewPager = _viewPager;
-        
-        _viewPager.changeIndex = ^(NSUInteger index) {
-            weakTitleBar.currentIndex = index;
-            for (UIButton *button in weakTitleBar.titleButtons) {
-                if (button.tag != index) {
-                    [button setTitleColor:[UIColor colorWithHex:0x909090] forState:UIControlStateNormal];
-                    button.transform = CGAffineTransformIdentity;
-                } else {
-                    [button setTitleColor:[UIColor colorWithHex:0x428bd1] forState:UIControlStateNormal];
-                    button.transform = CGAffineTransformMakeScale(1.2, 1.2);
-                }
-            }
-            [weakViewPager scrollToViewAtIndex:index];
-        };
-        
-        _viewPager.scrollView = ^(CGFloat offsetRatio, NSUInteger focusIndex, NSUInteger animationIndex) {
-            KLog(@"界面滚动中:viewpager is scrolling...");
-            UIButton *titleFrom = weakTitleBar.titleButtons[animationIndex];
-            UIButton *titleTo = weakTitleBar.titleButtons[focusIndex];
-            //CGFloat colorValue = (CGFloat)0x90 / (CGFloat)0xFF;
-            
-            [UIView transitionWithView:titleFrom duration:0.1 options:UIViewAnimationOptionTransitionCrossDissolve animations:^{
-                // 滚动时的颜色
-                //[titleFrom setTitleColor:[UIColor colorWithRed:colorValue*(1-offsetRatio) green:colorValue blue:colorValue*(1-offsetRatio) alpha:1.0]
-                //                forState:UIControlStateNormal];
-                titleFrom.transform = CGAffineTransformMakeScale(1 + 0.2 * offsetRatio, 1 + 0.2 * offsetRatio);
-            } completion:nil];
-            
-            [UIView transitionWithView:titleTo duration:0.1 options:UIViewAnimationOptionTransitionCrossDissolve animations:^{
-                //[titleTo setTitleColor:[UIColor colorWithRed:colorValue*offsetRatio green:colorValue blue:colorValue*offsetRatio alpha:1.0]
-                //              forState:UIControlStateNormal];
-                titleTo.transform = CGAffineTransformMakeScale(1 + 0.2 * (1-offsetRatio), 1 + 0.2 * (1-offsetRatio));
-            } completion:nil];
-        };
-        
-        _titleBar.titleButtonClicked = ^(NSUInteger index) {
-            [weakViewPager scrollToViewAtIndex:index];
-        };
-    }
-    
-    return self;
-}
-
-- (void)viewDidLoad {
-    [super viewDidLoad];
-    // Do any additional setup after loading the view.
-}
-
-- (void)didReceiveMemoryWarning {
-    [super didReceiveMemoryWarning];
-    // Dispose of any resources that can be recreated.
-}
-
-
-
-@end
