+++ WBSBlog/WWTagsCloudView.h
@@ -14,6 +14,8 @@
 #import <UIKit/UIKit.h>
 @interface WWTagsCloudView : UIView
 @property (strong, nonatomic) id<WWTagsCloudViewDelegate> delegate;
+
 -(id)initWithFrame:(CGRect)frame andTags:(NSArray*)tags andTagColors:(NSArray*)tagColors andFonts:(NSArray*)fonts andParallaxRate:(CGFloat)parallaxRate andNumOfLine:(NSInteger)lineNum;
+
 -(void)reloadAllTags;
 @end
